# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: DCS Back API
  description: |
    Backend for Dead Camera Studios. Three subsystems:
    
    **1. Images** — filesystem-based image upload/serve/delete (legacy).
    **2. Studio** — BytePlus ModelArk client (Seedance video, Seedream image, Assets API).
    **3. Files + Characters** — DB-backed file management (images, videos, audio) with character entity and file relations.
  version: 1.0.0
  contact:
    name: Dead Camera Studios
servers:
  - url: http://localhost:9099/api/v1
    description: Development
  - url: https://back.road2theoscars.tech/api/v1
    description: Production
paths:
...
  /projects/{id}/scenes:
    post:
      tags: [Projects]
      summary: Create a scene in a project
      operationId: createScene
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSceneRequest"
      responses:
        "201":
          description: Scene created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scene"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [Projects]
      summary: List scenes for a project
      operationId: listScenes
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Scene list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Scene"
                  success: { type: boolean }
                  message: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/scenes/{sceneId}:
    get:
      tags: [Projects]
      summary: Get scene with takes
      operationId: getScene
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: sceneId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Scene with takes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SceneWithTakes"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Projects]
      summary: Update scene fields
      operationId: updateScene
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: sceneId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSceneRequest"
      responses:
        "200":
          description: Scene updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scene"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Projects]
      summary: Soft delete a scene
      operationId: deleteScene
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: sceneId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          $ref: "#/components/responses/MessageOK"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/scenes/{sceneId}/takes:
    post:
      tags: [Projects]
      summary: Create a take in a scene
      operationId: createTake
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: sceneId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTakeRequest"
      responses:
        "201":
          description: Take created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Take"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [Projects]
      summary: List takes for a scene
      operationId: listTakes
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: sceneId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Take list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Take"
                  success: { type: boolean }
                  message: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/scenes/{sceneId}/takes/{takeId}:
    get:
      tags: [Projects]
      summary: Get a take by ID
      operationId: getTake
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: sceneId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: takeId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Take details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Take"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Projects]
      summary: Update take fields (video_url, status)
      operationId: updateTake
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: sceneId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: takeId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTakeRequest"
      responses:
        "200":
          description: Take updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Take"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Projects]
      summary: Soft delete a take
      operationId: deleteTake
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: sceneId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: takeId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          $ref: "#/components/responses/MessageOK"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
  schemas:
...
    ApiResponse:
      type: object
      properties:
        data: {}
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "success"
    ApiErrorResponse:
      type: object
      properties:
        data: { nullable: true, example: null }
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "error description"

    # --- Projects -------------------------------------------------
...
    CreateSceneRequest:
      type: object
      required: [number]
      properties:
        number:      { type: integer }
        name:        { type: string }
        description: { type: string }
    UpdateSceneRequest:
      type: object
      properties:
        number:      { type: integer }
        name:        { type: string }
        description: { type: string }
    Scene:
      type: object
      properties:
        id:          { type: string, format: uuid }
        project_id:  { type: string, format: uuid }
        number:      { type: integer }
        name:        { type: string }
        description: { type: string }
        active:      { type: boolean, default: true }
        created_at:  { type: string, format: date-time }
        updated_at:  { type: string, format: date-time }
        deleted_at:  { type: string, format: date-time, nullable: true }
    SceneWithTakes:
      type: object
      properties:
        scene: { $ref: "#/components/schemas/Scene" }
        takes:
          type: array
          items:
            $ref: "#/components/schemas/Take"
    CreateTakeRequest:
      type: object
      required: [number]
      properties:
        number: { type: integer, minimum: 1, maximum: 100 }
        status: { type: string, default: "pending" }
    UpdateTakeRequest:
      type: object
      properties:
        video_url:       { type: string }
        video_local_url: { type: string }
        status:          { type: string }
    Take:
      type: object
      properties:
        id:              { type: string, format: uuid }
        scene_id:        { type: string, format: uuid }
        number:          { type: integer }
        video_url:       { type: string }
        video_local_url: { type: string }
        status:          { type: string, enum: [pending, processing, completed, failed] }
        active:          { type: boolean, default: true }
        created_at:      { type: string, format: date-time }
        updated_at:      { type: string, format: date-time }
        deleted_at:      { type: string, format: date-time, nullable: true }
  responses:
    BadRequest:
      description: Bad request — missing or invalid fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
    MessageOK:
      description: Success
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse"
...